Addet Basic Turret with animation
Added Turret mesh without textures for now
Added animation controller script to play animation ingame
To play the animation hit play and ingame hit "Q", it will loop for now.
To turn this off click on the Mesh itself in the project explorer. In the inspector you will see a "loop button"

Mesh has a "root" for the turret itself for ez rotation transformation.